requirements.txt
absl-py==0.11.0
aiohttp==3.6.2
alpha-vantage==2.1.2
appdirs==1.4.3
argcomplete==1.11.1
astor==0.8.1
astroid==2.2.5
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs==20.3.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.1
bleach==1.5.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorlog==4.1.0
configparser==5.0.2
contextlib2==0.6.0.post1
cssselect==1.1.0
cvxopt==1.2.3
cycler==0.10.0
dataframe-image==0.1.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dm-sonnet==1.32
docker-pycreds==0.4.0
efficientnet==1.0.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
flask
fsspec==2021.4.0
future==0.18.2
gast==0.3.3
gdown==3.12.2
GetOldTweets3==0.0.11
gitdb==4.0.7
GitPython==3.1.14
google-pasta==0.2.0
gym==0.17.1
html5lib==0.9999999
idna==2.8
idna-ssl==1.1.0
image-classifiers==1.0.0
imageio==2.9.0
importlib-metadata==3.4.0
importlib-resources==1.4.0
iniconfig==1.1.1
ipykernel==5.1.4
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.20
itsdangerous==1.1.0
jax==0.1.61
jaxlib==0.1.42
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.5
jsonschema==3.2.0
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.1
llvmlite==0.34.0
lshashpy3==0.0.8
lxml==4.5.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
multidict==4.7.6
multitasking==0.0.9
nashpy==0.0.19
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.5
nltk==3.5
notebook==6.0.3
nox==2019.11.9
numpy==1.19.5
nvidia-pyindex==1.0.8
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
opt-einsum==3.3.0
packaging==20.9
pandas==0.25.3
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.6.2
pathlib==1.0.1
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip-autoremove==0.9.1
pluggy==0.13.1
ply==3.11
progressbar==2.5
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==1.0.18
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycore==17.4.17
pyftpdlib==1.5.5
pyglet==1.5.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.3.1
pymeanshift @ git+git://github.com/fjean/pymeanshift.git@1ba90da647342184ea7df378d7f21eba257a51d9
pyparsing==2.4.7
pyquery==1.4.1
pyrsistent==0.15.7
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-unibit==1.0.2
pytorch-lightning==1.1.8
pytorch-lightning-bolts==0.3.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.9.27
requests==2.22.0
requests-oauthlib==1.3.0
scikit-image==0.17.2
scikit-learn
scipy==1.5.4
segmentation-models==1.0.1
semantic-version==2.8.4
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.0.0
shortuuid==1.0.1
simplegeneric==0.8.1
six==1.15.0
smmap==4.0.0
soupsieve==2.0.1
split-folders==0.4.3
subprocess32==3.5.4
tensorflow==2.4.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
testresources==2.0.1
tifffile==2020.9.3
toml==0.10.2
torch==1.7.1
torchaudio==0.7.2
torchmetrics==0.3.1
torchtext==0.6.0
torchvision==0.8.2
tornado==6.0.4
tqdm==4.53.0
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.7
virtualenv==20.0.13
wandb==0.10.12
watchdog==2.0.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.0
yfinance==0.1.52
zipp==3.4.0